# [AIR][CI] Speed up HF CI by ~20%