--- a/python/ray/train/tests/test_huggingface_trainer.py
+++ b/python/ray/train/tests/test_huggingface_trainer.py
@@ -35,8 +35,8 @@ prompts = pd.DataFrame(
 
 # We are only testing Casual Language Modelling here
 
-model_checkpoint = "sshleifer/tiny-gpt2"
-tokenizer_checkpoint = "sgugger/gpt2-like-tokenizer"
+model_checkpoint = "hf-internal-testing/tiny-random-gpt2"
+tokenizer_checkpoint = "hf-internal-testing/tiny-random-gpt2"
 
 
 @pytest.fixture