Signed-off-by: Antoni Baum <antoni.baum@protonmail.com>

<!-- Thank you for your contribution! Please review https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR. If you don't have the access to it, we will shortly find a reviewer and assign them to your PR. -->

## Why are these changes needed?

Speeds up `HuggingFaceTrainer`/`Predictor` tests in CI by around ~20% by switching to a different GPT model. This is the same model Hugging Face team uses for their own CI.

Old model:
```
53.96s call     ray/train/tests/test_huggingface_trainer.py::test_e2e[epoch]
50.56s call     ray/train/tests/test_huggingface_trainer.py::test_e2e[no]
3.81s setup    ray/train/tests/test_huggingface_trainer.py::test_e2e[epoch]
3.74s setup    ray/train/tests/test_huggingface_trainer.py::test_e2e[no]
2.45s teardown ray/train/tests/test_huggingface_trainer.py::test_e2e[no]
2.27s teardown ray/train/tests/test_huggingface_trainer.py::test_e2e[epoch]
```

New model:
```
42.67s call     ray/train/tests/test_huggingface_trainer.py::test_e2e[epoch]
41.68s call     ray/train/tests/test_huggingface_trainer.py::test_e2e[no]
3.84s setup    ray/train/tests/test_huggingface_trainer.py::test_e2e[epoch]
3.69s setup    ray/train/tests/test_huggingface_trainer.py::test_e2e[no]
2.37s teardown ray/train/tests/test_huggingface_trainer.py::test_e2e[epoch]
2.36s teardown ray/train/tests/test_huggingface_trainer.py::test_e2e[no]
```

<!-- Please give a short summary of the change and the problem this solves. -->

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for https://docs.ray.io/en/master/.
- [ ] I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(