@Yard1 2022-08-31
@matthewdeng `pytest test.py --durations=0` - https://stackoverflow.com/questions/27884404/printing-test-execution-times-and-pinning-down-slow-tests-with-py-test